+++ packages/themes/_shared/CHANGELOG.md
@@ -1,8 +1,8 @@
-# @perch/themes-shared
+# @perch-app/themes-shared
 
 ## 0.0.1
 
 ### Patch Changes
 
 - Updated dependencies [5460338]
-  - @perch/core@0.1.0
+  - @perch-app/core@0.1.0